refactor: extract status-bubble keyboard rendering into status_bubble.py
Move build_status_keyboard() and its CB_STATUS_* imports out of
message_queue.py into a new status_bubble.py. message_queue.py was
mixing delivery logic with UI layout — this separation lets both modules
evolve independently. callback_data.py imports updated accordingly.
